requirements.txt
django==5.1.1
paho-mqtt==2.1.0
channels==4.1.0
daphne==4.1.2